+++ R/reexports.R
@@ -1,3 +1,8 @@
 #' @importFrom generics tidy
 #' @export
 generics::tidy
+
+#' @importFrom ggplot2 autoplot
+#' @export
+ggplot2::autoplot
+